exportar v1

+++ src/app/shared/modals/alert-edit/alert-edit.component.ts
@@ -2,7 +2,7 @@ import { Component, Input } from '@angular/core';
 import { ButtonComponent } from 'src/app/shared/components/button/button.component';
 import { IField, InputComponent } from 'src/app/shared/components/input/input.component';
 
-export interface IAlertEdit { title: string, message?: string, input?: IField, accept: (data?: any) => void, cancel: () => void,  delete: () => void}
+export interface IAlertEdit { title: string, message?: string, input?: IField, accept: (data?: any) => void, cancel: () => void,  delete?: () => void}
 
 @Component({
   selector: 'alert-edit',
@@ -31,7 +31,9 @@ export class AlertEditComponent {
     if (!this.data?.input) {
       return
     }
-    this.data?.delete()
+    if (this.data?.delete) {
+      this.data?.delete()
+    }
   }
 
 }
